requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# AWS integrations
boto3>=1.34.0
botocore>=1.34.0

# Deepgram
deepgram-sdk>=3.2.0

# HTTP requests for Merge and other APIs
requests>=2.31.0

# Data handling
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-dateutil>=2.8.2
